rename wind.py to winds.py and convert docstrings from reST to markdown
