+++ tests/test_dynamic_wind.py
@@ -18,7 +18,7 @@ from aleff import (
     wind,
     Ref,
 )
-from aleff._multishot.v1.wind import _get_wind_stack  # pyright: ignore[reportPrivateUsage]
+from aleff._multishot.v1.winds import _get_wind_stack  # pyright: ignore[reportPrivateUsage]
 
 
 # ---------------------------------------------------------------------------
